Release 0.15.0. See release notes or expand commit message for details.
[Breaking]
Inherited and implemented breaking changes from OwlCore.Nomad 0.10.0 and OwlCore.Nomad.Storage 0.10.0.
Code has been updated to use Cid and DagCid for TImmutablePointer and TMutablePointer, respectively. This is a breaking change for all types that implement these interfaces.
Existing event streams will need to be recreated or migrated to the new format, if any.